requirements.txt
python-dateutil==2.8.1
pymongo==3.11.0
fastapi==0.61.1
uvicorn==0.11.8
google-auth==1.21.3
python-jose==3.2.0
cryptography==3.1.1